pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "metadata-extraction-and-image-analysis-system"
version = "1.0.0"
description = "METADATA EXTRACTION AND IMAGE ANALYSIS SYSTEM."
readme = "README.md"
requires-python = ">=3.8"
license = { text = "MIT" }
authors = [
    { name = "Sivamani" }
]
keywords = [
    "forensics",
    "digital-forensics",
    "image-analysis",
    "metadata",
    "exif",
    "ai",
    "provenance",
    "security",
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Topic :: Security",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
dependencies = [
    "PyYAML>=6.0.1",
    "Pillow>=10.0.0",
    "pillow-heif>=0.10.0",
    "exifread>=3.0.0",
    "piexif>=1.1.3",
    "numpy>=1.24.3",
    "pandas>=2.0.3",
    "requests>=2.28.0",
    "tqdm>=4.65.0",
    "reportlab>=4.0.4",
    "scikit-learn>=1.3.0",
    "joblib>=1.3.1",
    "ImageHash>=4.2.0",
    "Jinja2>=3.1.2",
    "json5>=0.9.14",
    "rich>=13.5.2",
    "click>=8.1.7",
    "prompt-toolkit>=3.0.39",
    "pygments>=2.15.1",
    "python-dateutil>=2.8.2",
    "pytz>=2023.3",
    "tzlocal>=5.0.1",
    "psutil>=5.9.0",
]

[project.optional-dependencies]
analysis = [
    "matplotlib>=3.7.0",
    "seaborn>=0.12.0",
]
full = [
    "matplotlib>=3.7.0",
    "seaborn>=0.12.0",
    "opencv-python>=4.8.0.74",
    "scipy>=1.11.0",
    "networkx>=3.0",
]
gpu = [
    "torch>=2.0.0",
    "torchvision>=0.15.0",
]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "black>=23.7.0",
    "isort>=5.12.0",
    "flake8>=6.0.0",
    "mypy>=1.5.0",
]

[project.scripts]
metadata-extraction-and-image-analysis-system = "src.main:main"

[tool.setuptools]
include-package-data = true
license-files = ["LICENSE"]

[tool.setuptools.packages.find]
where = ["."]
include = ["src", "src.*"]
exclude = ["tests*", "scripts*", "datasets*", "results*"]

[tool.black]
line-length = 88
target-version = ["py38"]
include = "\\.pyi?$"
extend-exclude = """
/(
  \\.eggs
  | \\.hg
  | \\.mypy_cache
  | \\.tox
  | \\.venv
  | build
  | dist
  | results
  | datasets
  | forensic_results
  | final_narrative_results
)/
"""

[tool.isort]
profile = "black"
line_length = 88
known_first_party = ["src"]

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true
check_untyped_defs = true

[[tool.mypy.overrides]]
module = [
    "PIL.*",
    "cv2",
    "sklearn.*",
]
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --tb=short --strict-markers"
norecursedirs = ["tests/manual"]
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: integration tests",
    "unit: unit tests",
    "forensic: forensic-specific tests",
]

[tool.coverage.run]
source = ["src"]
omit = [
    "*/tests/*",
    "*/test_*",
    "*/__pycache__/*",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "@(abc\\.)?abstractmethod",
]

[tool.flake8]
max-line-length = 88
extend-ignore = ["E203", "W503"]
exclude = [
    "__pycache__",
    "build",
    "dist",
    "results",
    "datasets",
]
per-file-ignores = ["__init__.py:F401"]
